# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bidragskoll.se API",
    "description": "Funding discovery, eligibility, application and case-management platform.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/internal/cron/{job}": {
      "get": {
        "tags": [
          "internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 40
            },
            "in": "path",
            "name": "job",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "post": {
        "tags": [
          "internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "maxLength": 40
            },
            "in": "path",
            "name": "job",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/internal/readiness": {
      "get": {
        "tags": [
          "internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "in": "query",
            "name": "probe",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/feedback": {
      "post": {
        "tags": [
          "feedback"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "category",
                  "page",
                  "message"
                ],
                "properties": {
                  "category": {
                    "type": "string",
                    "enum": [
                      "facts",
                      "language",
                      "navigation",
                      "missing",
                      "technical",
                      "other"
                    ]
                  },
                  "page": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "opportunitySlug": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "message": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 4000
                  },
                  "locale": {
                    "type": "string",
                    "maxLength": 8
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/v1/events": {
      "post": {
        "tags": [
          "feedback"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "maxLength": 40
                  },
                  "props": {
                    "type": "object",
                    "additionalProperties": {
                      "type": [
                        "string",
                        "number",
                        "boolean"
                      ]
                    },
                    "maxProperties": 8
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}